requirements.txt
absl-py==2.4.0
affine==2.4.0
astunparse==1.6.3
attrs==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
ImageIO==2.37.2
joblib==1.5.3
keras==3.13.1
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==6.33.4
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
rasterio==1.5.0
requests==2.32.5
rich==14.3.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.1.28
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.0.1
